+++ clikt/src/commonTest/kotlin/com/github/ajalt/clikt/core/ContextTest.kt
@@ -1,5 +1,8 @@
 package com.github.ajalt.clikt.core
 
+import com.github.ajalt.clikt.parameters.arguments.argument
+import com.github.ajalt.clikt.parameters.arguments.multiple
+import com.github.ajalt.clikt.parameters.options.option
 import com.github.ajalt.clikt.testing.TestCommand
 import com.github.ajalt.clikt.testing.parse
 import io.kotest.assertions.throwables.shouldThrow
@@ -96,5 +99,21 @@ class ContextTest {
         shouldThrow<NoSuchOption> {
             C().context { helpOptionNames = setOf("--x") }.parse("--help")
         }
+    }
+
+    @Test
+    fun originalArgv() {
+        class C : TestCommand()
+        class S: TestCommand() {
+            val opt by option()
+            val args by argument().multiple()
+            override fun run_() {
+                opt shouldBe "o"
+                args shouldBe listOf("1", "2")
+                currentContext.originalArgv shouldBe listOf("s", "--opt", "o", "1", "2")
+            }
+        }
+
+        C().subcommands(S()).parse("s --opt o 1 2")
     }
 }